pyproject.toml
[tool.poetry]
name = "flask-template"
version = "0.1.0"
description = ""
authors = ["Edison Pama <45029122+edisonpama@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-restful = "^0.3.7"
tinydb = "^3.15"
PyStaticConfiguration = "^0.10.4"
pyyaml = "^5.2"
flask-jwt-extended = "^3.24"
blinker = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"